+++ src/app/models/url-navigate-options.model.ts
@@ -3,4 +3,5 @@ import { NavigationExtras } from '@angular/router';
 export interface UrlNavigateOptions {
   absolutePath?: string;
   navigationExtras?: NavigationExtras;
+  fragment?: string;
 }