+++ projects/fhi-angular-highcharts/CHANGELOG.md
@@ -1,8 +1,8 @@
 # Unreleased
 
-> Feb 15, 2024
+> Feb 16, 2024
 
-* :tada: **Enhancement** Add placeholder for metadata link
+* :tada: **Enhancement** Add link to metadata
 * :tada: **Enhancement** Change modal implementation from local component to a component in @folkehelseinstituttet/angular-components
 * :tada: **Enhancement** Removed table specific CSS classes, being replaced by general rules in @folkehelseinstituttet/style